# docs/frontend/openapi.json
{
  "components": {
    "schemas": {
      "AuthenticationResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "AuthenticationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginBody",
        "type": "object"
      },
...
      "RegisterBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterBody",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "role"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserStatus": {
        "enum": [
          "active",
          "disabled"
        ],
        "title": "UserStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "VIA Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Login rate limit exceeded."
          }
        },
        "summary": "Login",
        "tags": [
          "identity-access"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "__Secure-via_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Secure-Via Refresh"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Origin is not trusted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "identity-access"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "auth_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid or missing bearer token."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Me",
        "tags": [
          "identity-access"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "auth_refresh",
        "parameters": [
          {
            "in": "cookie",
            "name": "__Secure-via_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Secure-Via Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid or missing refresh cookie."
          },
          "403": {
            "description": "Origin is not trusted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Refresh rate limit exceeded."
          }
        },
        "summary": "Refresh",
        "tags": [
          "identity-access"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "auth_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "An account with this email already exists."
          },
          "422": {
            "description": "Invalid email or password."
          },
          "429": {
            "description": "Registration rate limit exceeded."
          }
        },
        "summary": "Register",
        "tags": [
          "identity-access"
        ]
      }
    },
...
  }
}